+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix for setting hint, remove the mention of resource consumption
 - Add some missing awaits on adding concentration effect and check item card
+- Add the source of concentration (i.e. spell's name) as flavor text for the check item card
 
 # 0.5